requirements.txt
# 트레이딩 봇 의존성 패키지 (Python 3.12 호환)

# 업비트 API
pyupbit==0.2.31

# 데이터 분석
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2

# 웹 프레임워크
flask>=2.3.0
flask-cors>=4.0.0

# 텔레그램 봇
python-telegram-bot>=20.7

# 시스템 모니터링
psutil>=5.9.0

# 환경 변수
python-dotenv>=1.0.0

# 데이터베이스
sqlalchemy>=2.0.0

# 로깅
colorlog>=6.7.0

# 수학/통계
scipy>=1.10.0
scikit-learn>=1.3.0

# 설정 파일
pyyaml>=6.0

# HTTP 클라이언트
requests>=2.31.0

# 스케줄링
schedule>=1.2.0

# 차트 생성
plotly>=5.15.0

# 암호화
cryptography>=41.0.0

# JSON 처리
jsonschema>=4.17.0

# 시간대 처리
pytz>=2023.3